Unit Creation Modifiers
Buildings now use stackable modifiers to indicate unit creation, upon
selecting to create a multiple unit the duration count is set to the
remaining time on its first cycle.

These modifiers seem to reset the building model, so to counter this on
BattleBeyond_updateModifier( event ) SetModel() is called to prevent the
building "disappearing".